% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyreader-package.R
\docType{package}
\name{easyreader-package}
\alias{easyreader}
\alias{easyreader-package}
\title{easyreader: Import a Dataset into R}
\description{
Import data from Excel, text files, and 
  statistical packages.
}
\keyword{internal}
