% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import data into a data frame}
\usage{
import(file, ...)
}
\arguments{
\item{file}{datafile to import}

\item{...}{parameters passed to the importing function}
}
\value{
data frame
}
\description{
Import data from Excel, text files, and
statistical packages.
}
\details{
The \code{import} function is a wrapper for functions in the
haven, readxl, and vroom packages.
}
\examples{
\dontrun{
if(interactive()){
 mydataframe <- import("mydata.csv")
 }
}
}
\seealso{
\code{\link[tools]{fileutils}}
 \code{\link[haven]{read_sas}},\code{\link[haven]{read_dta}},\code{\link[haven]{read_spss}}
 \code{\link[readxl]{read_excel}}
 \code{\link[vroom]{vroom}}
}
